pyproject.toml
[tool.poetry]
name = "quick-evac-backend"
version = "0.1.0"
description = "Backend for Quick Evacuation Management System"
authors = ["Your Name <your.email@example.com>"]

# Tell Poetry where to find your package
packages = [{ include = "backend", from = "." }]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
flask-sqlalchemy = "^2.5.1"
sqlalchemy = "^1.4.46"  # Specify a compatible version
flask-cors = "^3.0.10"
python-dotenv = "^0.19.1"
requests = "^2.26.0"
twilio = "^7.8.0"
googlemaps = "^4.6.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"